fix(sizing): 🐛 use native filesystem counter types (#77)
## 🚀 What

Keep filesystem capacity detection lint-clean on supported 64-bit
targets.

## 🤔 Why

`statvfs` already exposes block counts and fragment sizes as `u64` on
the supported Linux and Apple Silicon targets. Converting those values
back to `u64` triggers deny-level Clippy errors.

## 🔧 How

Multiply the native counters directly while retaining checked overflow
handling.

## ✅ Validation

- `cargo fmt --manifest-path Cargo.toml --all --check`
- `cargo clippy --locked -- -D warnings`

+++ server/src/sizing.rs
@@ -524,8 +524,9 @@ pub(crate) fn filesystem_available_bytes(path: &Path) -> Result<u64> {
     } else {
         stats.f_frsize
     };
-    u64::from(stats.f_bavail)
-        .checked_mul(u64::from(fragment_bytes))
+    stats
+        .f_bavail
+        .checked_mul(fragment_bytes)
         .ok_or_else(|| KvError::InvalidConfig("available filesystem size overflowed".into()))
 }
